Orb can change color upon Get request
Pass in the ID and it can change colors.
Also added user secret for orb ID
