+++ src/orby/appSettings.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace orby
+{
+    public class AppSettings
+    {
+        public string OrbID { get; set; }
+    }
+}
