+++ src/orby/appsettings.json
@@ -1,4 +1,7 @@
 ﻿{
+    "AppSettings": {
+      "OrbID": "Externalkey"
+    },
   "Logging": {
     "IncludeScopes": false,
     "LogLevel": {
